requirements.txt
selenium==4.10.0
python-dotenv==0.21.0
pandas==1.5.3
matplotlib==3.5.3
reportlab==3.6.12
openpyxl==3.0.10 